+++ dns-resolver/src/DNSQuestion.hpp
@@ -5,13 +5,13 @@
 
 class DNSQuestion {
 private:
-    int type;
-    int klass;
+    uint16_t type;
+    uint16_t klass;
     std::vector<uint8_t> name;
     
 public:
     DNSQuestion() {};
-    DNSQuestion(int type, int klass, std::vector<uint8_t> name);
+    DNSQuestion(uint16_t type, uint16_t klass, std::vector<uint8_t> name);
     std::vector<uint8_t> ToBytes();
 };
